[vtadmin-web] Add tabs to Workflow view
